+++ backend/typescript/rest/entityRoutes.ts
@@ -1,74 +1,88 @@
-import { Router } from "express";
-import EntityService from "../services/implementations/EntityService";
+import { Router } from 'express';
+import { Entity, IEntity } from "../models/entity.mgmodel"
+import EntityService, { EntityRequestDTO } from "../services/EntityService";
 
-const entityRouter: Router = Router();
-const entService = new EntityService();
+export const entityRouter: Router = Router();
 
 /* Create entity Object */
-entityRouter.post("/", async (req, res) => {
-  try {
-    const newEntity = await entService.createEntity({
-      stringField: req.body.stringField,
-      intField: req.body.intField,
-      enumField: req.body.enumField,
-      stringArrayField: req.body.stringArrayField,
-      boolField: req.body.boolField,
-    });
-    res.status(201).json(newEntity);
-  } catch (e) {
-    res.status(500).send(e.message);
-  }
+entityRouter.post("/entities/", async (_req, res) => {
+    const entityObj: EntityRequestDTO = _req.body.entity;
+    const IentityObj = {} as IEntity;
+    const entService = new EntityService();
+    IentityObj.string_field = entityObj.string_field;
+    IentityObj.int_field = entityObj.int_field;
+    IentityObj.enum_field = entityObj.enum_field;
+    IentityObj.string_array_field = entityObj.string_array_field;
+    IentityObj.bool_field = entityObj.bool_field;
+
+    try {
+        await entService.createEntity(IentityObj);
+        res.send('Success')
+    }
+    catch (e) {
+        res.status(500).send(e.message);
+    }
 });
 
+
 /* Get all entity objects */
-entityRouter.get("/", async (req, res) => {
-  try {
-    const entities = await entService.getEntities();
-    res.status(200).json(entities);
-  } catch (e) {
-    res.status(500).send(e.message);
-  }
+entityRouter.get("/entities", async (_req, res) => {
+    const entService = new EntityService();
+
+    try {
+        const result = await entService.getEntities();
+        res.send(result)
+    }
+    catch (e) {
+        res.status(500).send(e.message);
+    }
 });
 
 /* Get entity object by id */
-entityRouter.get("/:id", async (req, res) => {
-  const { id } = req.params;
+entityRouter.get("/entities/:id", async (_req, res) => {
+    const entService = new EntityService();
+    const id: string = _req.params.id;
 
-  try {
-    const entity = await entService.getEntity(id);
-    res.status(200).json(entity);
-  } catch (e) {
-    res.status(500).send(e.message);
-  }
+    try {
+        const result = await entService.getEntity(id);
+        res.send(result)
+    }
+    catch (e) {
+        res.status(500).send(e.message);
+    }
 });
 
 /* Update entity object by id */
-entityRouter.put("/:id", async (req, res) => {
-  const { id } = req.params;
-  try {
-    const entity = await entService.updateEntity(id, {
-      stringField: req.body.stringField,
-      intField: req.body.intField,
-      enumField: req.body.enumField,
-      stringArrayField: req.body.stringArrayField,
-      boolField: req.body.boolField,
-    });
-    res.status(200).json(entity);
-  } catch (e) {
-    res.status(500).send(e.message);
-  }
+entityRouter.put("/entities/:id", async (_req, res) => {
+    const entService = new EntityService();
+    const id: string = _req.params.id;
+    const entityObj: EntityRequestDTO = _req.body.entity;
+
+    const IentityObj = {} as IEntity;
+    IentityObj.string_field = entityObj.string_field;
+    IentityObj.int_field = entityObj.int_field;
+    IentityObj.enum_field = entityObj.enum_field;
+    IentityObj.string_array_field = entityObj.string_array_field;
+    IentityObj.bool_field = entityObj.bool_field;
+    try {
+        const result = await entService.updateEntity(id, IentityObj);
+        res.send(result)
+    }
+    catch (e) {
+        res.status(500).send(e.message);
+    }
 });
 
 /* Delete entity object by id */
-entityRouter.delete("/:id", async (req, res) => {
-  const { id } = req.params;
+entityRouter.delete("/entities/:id", async (_req, res) => {
+    const entService = new EntityService();
+    const id: string = _req.params.id;
 
-  try {
-    await entService.deleteEntity(id);
-    res.status(204).send();
-  } catch (e) {
-    res.status(500).send(e.message);
-  }
+    try {
+        entService.deleteEntity(id);
+        res.send('Success')
+    }
+    catch (e) {
+        res.status(500).send(e.message);
+    }
 });
-
-export default entityRouter;
